Update rotary_embedding_torch.py
When I was using a model that depended on this library, and I unfroze the last layer of the transformer I got this error:

RuntimeError: one of the variables needed for gradient computation has been modified by an in-place operation:

I don't think this function intends to modify the variable "t" in place.  So I propose this change to remove the in-place error.